pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ergodspy"
version = "0.1.0"
description = "Ergonomic DSPy signatures - write DSPy signatures with minimal boilerplate"
readme = "README.md"
authors = [{name = "Harish Batra", email = "hrbatra@utexas.edu"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["dspy", "llm", "language-models", "prompts", "ergonomic"]
dependencies = [
    "dspy-ai>=2.0.0",
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/hbatra21/ergodspy"
Repository = "https://github.com/hbatra21/ergodspy"
Issues = "https://github.com/hbatra21/ergodspy/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "ruff",
]

[tool.setuptools]
packages = ["ergodspy"]
